+++ Source/Model/OptionRow.swift
@@ -52,8 +52,11 @@ open class OptionRow<T: UITableViewCell>: OptionSelectable, Equatable {
   /// The state of selection.
   public var isSelected: Bool = false {
     didSet {
-      if isSelected != oldValue {
-        action?(self)
+      guard isSelected != oldValue else {
+        return
+      }
+      DispatchQueue.main.async {
+        self.action?(self)
       }
     }
   }